[core] implement soft delete

+++ src/TeamCloud.Data/CosmosDb/CosmosDbComponentRepository.cs
@@ -83,7 +83,10 @@ namespace TeamCloud.Data.CosmosDb
             return null;
         }
 
-        public override async IAsyncEnumerable<Component> ListAsync(string projectId)
+        public override IAsyncEnumerable<Component> ListAsync(string projectId)
+            => ListAsync(projectId, includeDeleted: false);
+
+        public async IAsyncEnumerable<Component> ListAsync(string projectId, bool includeDeleted)
         {
             if (projectId is null)
                 throw new ArgumentNullException(nameof(projectId));
@@ -95,6 +98,9 @@ namespace TeamCloud.Data.CosmosDb
                 .ConfigureAwait(false);
 
             var queryString = $"SELECT * FROM c WHERE c.projectId = '{projectIdParsed}'";
+
+            if (!includeDeleted)
+                queryString += " AND NOT IS_DEFINED(c.deleted)";
 
             var query = new QueryDefinition(queryString);
 
@@ -112,21 +118,41 @@ namespace TeamCloud.Data.CosmosDb
             }
         }
 
-        public override async Task<Component> RemoveAsync(Component component)
+        public override Task<Component> RemoveAsync(Component component)
+            => RemoveAsync(component, soft: true);
+
+        public async Task<Component> RemoveAsync(Component component, bool soft)
         {
             if (component is null)
                 throw new ArgumentNullException(nameof(component));
+
+            if (soft && component.Deleted.HasValue && component.TTL.HasValue)
+                return component;
 
             var container = await GetContainerAsync()
                 .ConfigureAwait(false);
 
             try
             {
-                var response = await container
-                    .DeleteItemAsync<Component>(component.Id, GetPartitionKey(component))
-                    .ConfigureAwait(false);
+                if (soft)
+                {
+                    component.Deleted = DateTime.UtcNow;
+                    component.TTL = GetSoftDeleteTTL();
 
-                return response.Resource;
+                    var response = await container
+                        .UpsertItemAsync(component, GetPartitionKey(component))
+                        .ConfigureAwait(false);
+
+                    return response.Resource;
+                }
+                else
+                {
+                    var response = await container
+                        .DeleteItemAsync<Component>(component.Id, GetPartitionKey(component))
+                        .ConfigureAwait(false);
+
+                    return response.Resource;
+                }
             }
             catch (CosmosException cosmosEx) when (cosmosEx.StatusCode == HttpStatusCode.NotFound)
             {
@@ -134,9 +160,9 @@ namespace TeamCloud.Data.CosmosDb
             }
         }
 
-        public async Task RemoveAllAsync(string projectId)
+        public async Task RemoveAllAsync(string projectId, bool soft)
         {
-            var components = ListAsync(projectId);
+            var components = ListAsync(projectId, includeDeleted: !soft);
 
             if (await components.AnyAsync().ConfigureAwait(false))
             {
@@ -146,8 +172,23 @@ namespace TeamCloud.Data.CosmosDb
                 var batch = container
                     .CreateTransactionalBatch(GetPartitionKey(projectId));
 
-                await foreach (var component in components.ConfigureAwait(false))
-                    batch = batch.DeleteItem(component.Id);
+                if (soft)
+                {
+                    var now = DateTime.UtcNow;
+                    var ttl = GetSoftDeleteTTL();
+
+                    await foreach (var component in components.ConfigureAwait(false))
+                    {
+                        component.Deleted = now;
+                        component.TTL = ttl;
+                        batch = batch.UpsertItem(component);
+                    }
+                }
+                else
+                {
+                    await foreach (var component in components.ConfigureAwait(false))
+                        batch = batch.DeleteItem(component.Id);
+                }
 
                 await batch
                     .ExecuteAsync()
@@ -155,17 +196,17 @@ namespace TeamCloud.Data.CosmosDb
             }
         }
 
-        public async Task RemoveAsync(string projectId, string id)
-        {
-            var component = await GetAsync(projectId, id)
-                .ConfigureAwait(false);
+        // public async Task RemoveAsync(string projectId, string id)
+        // {
+        //     var component = await GetAsync(projectId, id)
+        //         .ConfigureAwait(false);
 
-            if (component != null)
-            {
-                await RemoveAsync(component)
-                    .ConfigureAwait(false);
-            }
-        }
+        //     if (component != null)
+        //     {
+        //         await RemoveAsync(component)
+        //             .ConfigureAwait(false);
+        //     }
+        // }
 
         public override async Task<Component> SetAsync(Component component)
         {
